# openapi.json
{"openapi":"3.1.0","info":{"title":"API da Aula 3","description":"\n    API desenvolvida durante a aula 2, contendo endpoints de exemplo e soma\n\n    - /teste: retorna uma mensagem de sucesso\n    - /soma/numero1/numero2: recebe dois números e retorna a soma\n","termsOfService":"http://example.com/terms/","contact":{"name":"Rogério Rodrigues Carvalho","url":"http://github.com/rogerior/","email":"rogerior@ufg.br"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.html"},"version":"0.1"},"paths":{"/llm/gerar_historia":{"post":{"summary":"Gera uma história sobre o tema informado por parâmetro","description":"Gera uma história em português brasileiro sobre um tema específico usando a API Groq.","operationId":"gerar_historia_llm_gerar_historia_post","parameters":[{"name":"tema","in":"query","required":true,"schema":{"type":"string","title":"Tema"}},{"name":"api_token","in":"query","required":true,"schema":{"type":"integer","title":"Api Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/operacoes/teste":{"get":{"tags":["Teste"],"summary":"Retorna mensagem de teste","description":"Retorna uma mensagem de exemplo para testar e verificar se deu certo","operationId":"hello_world_operacoes_teste_get","parameters":[{"name":"api_token","in":"query","required":true,"schema":{"type":"integer","title":"Api Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/operacoes/soma/{numero1}/{numero2}":{"post":{"tags":["Operações matemáticas simples"],"summary":"Recebe dois números na url e retorna a soma","operationId":"soma_operacoes_soma__numero1___numero2__post","parameters":[{"name":"numero1","in":"path","required":true,"schema":{"type":"integer","title":"Numero1"}},{"name":"numero2","in":"path","required":true,"schema":{"type":"integer","title":"Numero2"}},{"name":"api_token","in":"query","required":true,"schema":{"type":"integer","title":"Api Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/operacoes/soma/v2":{"post":{"tags":["Operações matemáticas simples"],"summary":"Recebe dois números no corpo da requisição e retorna a soma","operationId":"soma_formato2_operacoes_soma_v2_post","parameters":[{"name":"numero1","in":"query","required":true,"schema":{"type":"integer","title":"Numero1"}},{"name":"numero2","in":"query","required":true,"schema":{"type":"integer","title":"Numero2"}},{"name":"api_token","in":"query","required":true,"schema":{"type":"integer","title":"Api Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Resultado"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/operacoes/soma/v3":{"post":{"tags":["Operações matemáticas simples"],"summary":"Recebe dois números no corpo da requisição e retorna a soma","operationId":"soma_formato3_operacoes_soma_v3_post","parameters":[{"name":"api_token","in":"query","required":true,"schema":{"type":"integer","title":"Api Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Numero"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Resultado"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/operacoes/divisao/{numero1}/{numero2}":{"post":{"tags":["Operações matemáticas simples"],"summary":"Recebe dois números na url e retorna a divisão","operationId":"divisao_operacoes_divisao__numero1___numero2__post","parameters":[{"name":"numero1","in":"path","required":true,"schema":{"type":"integer","title":"Numero1"}},{"name":"numero2","in":"path","required":true,"schema":{"type":"integer","title":"Numero2"}},{"name":"api_token","in":"query","required":true,"schema":{"type":"integer","title":"Api Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/operacoes/operacao":{"post":{"tags":["Operações matemáticas simples"],"summary":"Recebe dois números e o tipo de operação e retorna o resultado","operationId":"operacao_operacoes_operacao_post","parameters":[{"name":"tipo","in":"query","required":true,"schema":{"$ref":"#/components/schemas/TipoOperacao"}},{"name":"api_token","in":"query","required":true,"schema":{"type":"integer","title":"Api Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Numero"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Numero":{"properties":{"numero1":{"type":"integer","title":"Numero1"},"numero2":{"type":"integer","title":"Numero2"}},"type":"object","required":["numero1","numero2"],"title":"Numero","description":"Classe Numero que herda de BaseModel.\n\nAtributos:\n    numero1 (int): Primeiro número inteiro.\n    numero2 (int): Segundo número inteiro."},"Resultado":{"properties":{"resultado":{"type":"integer","title":"Resultado"}},"type":"object","required":["resultado"],"title":"Resultado","description":"Classe que representa o resultado de uma operação.\n\nAtributos:\n    resultado (int): O valor do resultado."},"TipoOperacao":{"type":"string","enum":["soma","subtracao","multiplicacao","divisao"],"title":"TipoOperacao","description":"Enumeração que representa os tipos de operações matemáticas.\n\nAtributos:\n    soma (str): Representa a operação de soma.\n    subtracao (str): Representa a operação de subtração.\n    multiplicacao (str): Representa a operação de multiplicação.\n    divisao (str): Representa a operação de divisão."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}